\name{creSGL-package}
\alias{creSGL-package}
\docType{package}
\title{Fit a GLM with a combination of lasso and overlapping group lasso regularization
}
\description{Fit a regularized generalized linear model via penalized maximum likelihood.  The model is fit for a path of values of the penalty parameter. Fits linear and logistic models.
}
\details{
\tabular{ll}{
Package: \tab creSGL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-20\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Only 4 functions:
\code{creSGL}
\code{cvSGL}
\code{predictSGL}
\code{plot.cvSGL}
}
\author{Kourosh Zarringhalam [aut] and David Degras [ctb]

Modified from SGL package: Noah Simon, Jerome Friedman, Trevor Hastie, and Rob Tibshirani

Maintainer: Kourosh Zarringhalam <kourosh.zarringhalam@umb.edu>
}
\references{Simon, N., Friedman, J., Hastie T., and Tibshirani, R. (2011)
  \emph{A Sparse-Group Lasso}, \cr
  \url{http://web.stanford.edu/~hastie/Papers/SGLpaper.pdf}
}
